io: Change size of alignment to 64bit unsigned
The alignment was defined as a unsigned machine word, but should allow
for 64bit values. The function to set the alignment explicitly already
took a 64bit parameter, so we store 64bit alignment values now.

+++ io/server/src/resource.cc
@@ -78,7 +78,7 @@ Resource::dump(char const *ty, int indent) const
   else
     printf("%*.s%-6s%c [%014llx-%014llx %llx]%s%s (align=%llx flags=%lx)\n",
            indent, "", ty, provided() ? '*' : ' ', _s, _e, (l4_uint64_t)size(),
-           tp ? " " : "", tp ? tp : "", (unsigned long long)alignment(), flags());
+           tp ? " " : "", tp ? tp : "", alignment(), flags());
 }
